pc collector: split code for 8/16/32 bit into separate files
